app/requirements.txt
Flask
gunicorn==21.2.0
google-cloud-monitoring==2.19.1
google-auth

